pgAdmin FIX, createsuperuser FIX and REST for Book
Fixed the problem of pgAdmin container, not correctly loading the .pgpass file from the container, added the PGPASS_FILE environment variable to the docker_compose.yml file. Fixed a problem with the createsuperuser instruction in the backend Dockerfile, which caused the unexpected exit from the container if the volume was not cleared before running again the application (error already exists that user). Then implemented a first backend response to REST requests for the Book model. In particular implemented the /books and the /books/isbn JSON responses, in order for the frontend to be able to retrieve information from the database through Django. Added some preloaded books, in the fixtures folder of the Django bookline app, which consists in a single json file with a bunch of data.