Several fixes throughout the application, and fix for reservation views, search function updated and inserted into the catalog
Solved the problem for which performing certain action brought to a page reload which landed for some moments on the login page before going on.
Added a spinner to give feedback to the user when a page is still loading before being presented on screen.
Solved issues with reservation views, in particular on availability, creation, cancel and return.
Search bar inserted in catalog and updated in its functioning.